+++ src/app/(auth)/register/page.tsx
@@ -1,0 +1,7 @@
+import AuthForm from "@components/auth/AuthForm";
+
+const Register = () => {
+  return <AuthForm isLogin={false} />;
+};
+
+export default Register;